refactor code to 2 components
